FEAT: Finalised project structure
