+++ pdm.lock
@@ -103,6 +103,17 @@ requires_python = ">=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*, !=3.4.*"
 summary = "library with cross-python path, ini-parsing, io, code, log facilities"
 
 [[package]]
+name = "pyhtmlproofer"
+version = "0.1.0"
+requires_python = ">=3.9"
+path = "."
+summary = "htmlproofer - A tool for validating internal & external links in HTML files / Websites"
+dependencies = [
+    "aiohttp",
+    "requests>=2.28.1",
+]
+
+[[package]]
 name = "pyparsing"
 version = "3.0.9"
 requires_python = ">=3.6.8"
@@ -160,7 +171,7 @@ dependencies = [
 
 [metadata]
 lock_version = "3.1"
-content_hash = "sha256:ccc04b011f62f469d6c758026b50a152ccb7b3d031f5548ca2a39e0ff517df15"
+content_hash = "sha256:35e9a9a000ee02b992f2681a02fdb9b37ea973a60e7ddff0268e0699c1fcc8f1"
 
 [metadata.files]
 "aiohttp 3.8.1" = [
