+++ pyHTMLProofer/Utils/__init__.py
@@ -1,0 +1,3 @@
+"""Initialise Utils module."""
+
+from .Log import Log
