+++ pyHTMLProofer/Version.py
@@ -1,0 +1,4 @@
+"""Version of pyHTMLProofer."""
+from __future__ import unicode_literals
+
+__version__ = "0.1.0.alpha"
