+++ tests/test_app.py
@@ -1,6 +1,8 @@
-from src.HTMLProofer.htmlproofer import HTMLProofer
+import pyHTMLProofer
 
 
-def test_htmlproofer():
-    proofer = HTMLProofer()
-    assert proofer.check_file("tests/test_app.py") == True
+# pyHTMLProofer.check_file("tests/out/test_dir1/test_file1.html")
+# pyHTMLProofer.check_directory("tests/out")
+# pyHTMLProofer.check_directories(["tests/out/test_dir1", "tests/out/test_dir2"])
+# pyHTMLProofer.check_links(["https://www.google.com", "https://www.google.com/"])
+pyHTMLProofer.check_sitemap("https://cloudbytes.dev/sitemap.xml", options={"log_level": "ERROR"})
